7.4

+++ controllers/videoController.js
@@ -51,7 +51,6 @@ export const videoDetail = async (req,res) =>{
 	try{
 	const video = await Video.findById(id).populate("creator");
 	//객체를 데려오는 함수. (objectID 타입에만 사용 가능)
-	console.log(video);
 	res.render("videoDetail", {  pageTitle : video.title, video });
 	}catch(error){
 	console.log(error);
@@ -65,11 +64,16 @@ export const getEditVideo = async (req,res) => {
 	const {
 		params: {id}
 	} = req;
-	try{
-		const video = await Video.findById(id);
-		res.render("editVideo", { pageTitle : `Edit ${video.title}`, video });
-	}catch(error){
+	 try {
+    const video = await Video.findById(id);
+    if (video.creator != req.user.id) {
+		throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    	}
+	} catch (error) {
 		res.redirect(routes.home);
+		console.log(error);
 	}
 }
 
@@ -92,8 +96,13 @@ export const deleteVideo = async (req,res) => {
 		params: {id}
 	} = req;
 	try{
-		await Video.findOneAndRemove({ _id: id });
-	}catch(error){
+		const video = await Video.findById(id);
+		if (video.creator != req.user.id) {
+			throw Error();
+		} else {
+			await Video.findOneAndRemove({_id:id});
+    	}
+	} catch (error) {
 		console.log(error);
 	}
 	res.redirect(routes.home);
